#8865: Move dispatch profiling scripts to ttnn subfolder
