+++ tests/ttnn/profiling/ops_for_profiling.py
