+++ tests/ttnn/profiling/reference.txt
@@ -364,7 +364,7 @@ tt_lib.tensor.threshold,40,4.85,4.93,5.5,5.0
 tt_lib.tensor.threshold_bw,40,7.04,7.22,7.62,7.08
 tt_lib.tensor.tilize,40,0.01,0.02,0.11,0.0
 tt_lib.tensor.tilize_with_val_padding,40,0.54,1.03,4.61,0.87
-tt_lib.tensor.tilize_with_zero_padding,40,0.01,0.02,0.11,0.0
+tt_lib.tensor.tilize_with_zero_padding,200,0.015,0.015,0.019
 tt_lib.tensor.transpose,40,0.88,0.95,1.23,0.99
 tt_lib.tensor.tril,24,99.64,99.76,99.54,1.65
 tt_lib.tensor.triu,24,99.97,100.18,99.82,1.2
